Fix diazo rules for search view
Since the search view uses ajax requests to update its contents the
content area filtering for template partials will not work.
Unfortunately a custom theme template is assigned to the search
templates body class which caused a styled result template in
return to the ajax request.

NOTE: Moving the notheme parameter on top fixes this issue but might
cause other unwanted sideeffects. This is a potentially dangerous
change.